drivers: lkss: lab2: ex2
Implement a char driver that writes a string to driver and
reads it back reversed.

This driver uses class_create and device_create to create a
device node automatically in /dev/, which:
- provides a user-visible device (/dev/reverse_char).
- registers the driver with the device model (via class_create,
device_create).

Compared to the previous simple driver:
- that one only called register_chrdev and relied on manual
creation of the device node (using mknod).
- this one is more complete and user-friendly,
creating both sysfs entries and the device file automatically.

Signed-off-by: Iuliana Prodan <iuliana.prodan@nxp.com>