gfs2: Avoid unnecessary transactions in evict_linked_inode
In evict_linked_inode(), the truncate_inode_pages() calls are carried
out inside a transaction.  This code was added to what was then function
gfs2_delete_inode() in commit 16615be18cadf ("[GFS2] Clean up journaled
data writing").

Those transactions are only used for creating revokes for the jdata
buffers in the journal, so don't create such transactions when we know
that the journal doesn't contain any jdata buffers for this inode.

Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>