drbd: add legacy 8.4 compatibility layer
Add compatibility layer for interoperability with DRBD 8.4
protocol and behavior.

This introduces the ability to read and write DRBD 8.4 metadata, as well
as restoring the old 8.4-style /proc/drbd format if a resource is in
"compatibility mode".
This mode is enabled when the corresponding option is sent from a
drbdsetup command.

Signed-off-by: Christoph Böhmwalder <christoph.boehmwalder@linbit.com>